backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
numpy==1.24.3
openai-whisper==20231117
websockets==12.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
sounddevice==0.4.6
torch==2.1.1
torchaudio==2.1.1
transformers==4.35.2
scipy==1.11.4
python-json-logger==2.0.7
structlog==23.2.0
rich==13.7.0
